+++ src/SimpleCliniq.Module.Core.Domain/Models/MRekanan.cs
@@ -7,7 +7,7 @@ namespace SimpleCliniq.Module.Core.Domain.Models;
 
 public partial class MRekanan
 {
-    public decimal IdRekanan { get; set; }
+    public int IdRekanan { get; set; }
 
     public string NmRekanan { get; set; }
